refactor(mfa): simplify UI-walk harness with xstate TestModel
Replace the hand-rolled path harness (wrapPath/assertMatchingStates/RawPath)
with xstate/graph's public TestModel, built via the constructor to skip the
validateMachine `after` check that createTestModel throws on. Coverage paths
still come from the standalone getShortestPaths (auto-synthesized events),
routed through TestModel.getPaths so new/actor-internal transitions stay
covered with no edit; getPathsFromEvents drives the teardown lap only.

- Drop the custom toStateValue: matchesState splits dot-path keys itself.
- Merge realUi/{renderModal,userGestures} -> harness.tsx and
  {jestMocks,mocks} -> mocks.ts.
- Mock @libs/XStateInspector as {inspect: undefined} and keep the real
  useInspectedMachine instead of reproducing the hook.
- Reuse the shared Navigation manual mock over a catch-all Proxy so a new
  unmocked Navigation dependency fails loudly.
- Name walk tests from driven event types, not TestModel's payload-dumping
  path.description.
- Remove the dead safe-area initialMetrics (the global mock ignores them).
