+++ tests/utils/mfa/reachableStates.ts
@@ -1,27 +1,22 @@
-import type {AnyStateNode, StateValue} from 'xstate';
+import type {AnyStateNode} from 'xstate';
 
-type SettleableLeafState = {description: string; value: StateValue};
-
-function toStateValue(path: string[]): StateValue {
-    if (path.length <= 1) {
-        return path.at(0) ?? '';
-    }
-    const [head, ...rest] = path;
-    return {[head]: toStateValue(rest)};
-}
+type SettleableLeafState = {description: string};
 
 function getSettleableLeafStates(node: AnyStateNode): SettleableLeafState[] {
     const children = Object.values(node.states);
     if (children.length > 0) {
         return children.flatMap(getSettleableLeafStates);
     }
-    // A leaf with an `always` transition is a transient router that leaves on entry and never settles.
+    // A leaf with an `always` transition is treated as a transient router that leaves on entry and never
+    // settles. This assumes the machine convention that `always` on a leaf is unconditional (the current
+    // `preparing` router is). A GUARDED `always` that can fail to fire would settle and should count as a
+    // leaf; revisit this (and cover parent-level `always`) before adding conditional routers.
     if ((node.always?.length ?? 0) > 0) {
         return [];
     }
-    return [{description: node.path.join('.'), value: toStateValue(node.path)}];
+    // The dot-path description doubles as a state-value key: `matchesState` splits it on `.` before comparing.
+    return [{description: node.path.join('.')}];
 }
 
 export default getSettleableLeafStates;
-export {toStateValue};
 export type {SettleableLeafState};
