+++ tests/utils/mfa/realUi/harness.tsx
@@ -1,10 +1,57 @@
-import {act, fireEvent, screen} from '@testing-library/react-native';
+import {act, fireEvent, render, screen} from '@testing-library/react-native';
+import React, {useEffect} from 'react';
+import ComposeProviders from '@components/ComposeProviders';
+import {LocaleContextProvider} from '@components/LocaleContextProvider';
+import {MultifactorAuthenticationContextProviders, useMultifactorAuthentication} from '@components/MultifactorAuthentication/Context';
 import type {MfaEvent} from '@components/MultifactorAuthentication/machine/types';
 import {handleInitialScreenLayout} from '@components/MultifactorAuthentication/mfaNavigation';
+import OnyxListItemProvider from '@components/OnyxListItemProvider';
+import MultifactorAuthenticationModalNavigator from '@navigation/AppNavigator/Navigators/MultifactorAuthenticationModalNavigator';
 import waitForBatchedUpdatesWithAct from '../../waitForBatchedUpdatesWithAct';
 import {MFA_TEST_SCENARIO_NAME} from '../flowFixtures';
 import {pendingModalClose} from './mocks';
-import {getMfaControls} from './renderModal';
+
+type MfaUiControls = {
+    executeScenario: ReturnType<typeof useMultifactorAuthentication>['executeScenario'];
+};
+
+const controlsHolder: {current: MfaUiControls | undefined} = {current: undefined};
+
+/**
+ * Renders nothing. It sits inside the providers only to capture the live context API so the event
+ * executors can start a flow through the public API.
+ */
+function MfaControlsCapture() {
+    const {executeScenario} = useMultifactorAuthentication();
+    useEffect(() => {
+        controlsHolder.current = {executeScenario};
+    });
+    return null;
+}
+
+/**
+ * Mounts the real provider stack and the real MFA modal navigator the app renders in production. No
+ * SafeAreaProvider: the global `react-native-safe-area-context` mock returns fixed insets/frame from
+ * plain hooks, so nothing in the tree needs a provider above it.
+ */
+function renderMfaUi() {
+    controlsHolder.current = undefined;
+    return render(
+        <ComposeProviders components={[OnyxListItemProvider, LocaleContextProvider]}>
+            <MultifactorAuthenticationContextProviders>
+                <MfaControlsCapture />
+                <MultifactorAuthenticationModalNavigator />
+            </MultifactorAuthenticationContextProviders>
+        </ComposeProviders>,
+    );
+}
+
+function getMfaControls(): MfaUiControls {
+    if (!controlsHolder.current) {
+        throw new Error('MFA UI controls were not captured. Call renderMfaUi() and await waitForBatchedUpdatesWithAct() first.');
+    }
+    return controlsHolder.current;
+}
 
 type MfaEventType = MfaEvent['type'];
 type MfaEventExecutor = () => Promise<void>;
@@ -44,4 +91,4 @@ const mfaEventExecutors = {
 } satisfies Record<MfaEventType, MfaEventExecutor>;
 /* eslint-enable @typescript-eslint/naming-convention */
 
-export default mfaEventExecutors;
+export {renderMfaUi, mfaEventExecutors};
